requirements/newreq.txt
Django==1.6.7
Fabric==1.6.0
Pillow==2.5.3
PyPDF2==1.23
South==0.7.6
Werkzeug==0.8.3
argparse==1.2.1
coverage==3.6
dj-database-url==0.2.1
django-easy-pdf==0.1.0
django-extensions==1.1.1
django-grappelli==2.5.3
django-nose==1.1
ecdsa==0.11
flake8==2.0
html5lib==0.999
ipython==0.13.2
mccabe==0.2.1
nose==1.3.0
nosexcover==1.0.8
paramiko==1.14.1
pep8==1.5.7
pycrypto==2.6.1
pyflakes==0.8.1
reportlab==2.7
six==1.7.3
sqlparse==0.1.11
wsgiref==0.1.2
xhtml2pdf==0.0.6
